+++ src/main/java/com/vwo/VWO.java
@@ -66,17 +66,17 @@ public class VWO extends VWOClient {
 
     public static VWO init(VWOInitOptions options) {
         if (options == null || options.getSdkKey() == null || options.getSdkKey().isEmpty()) {
-            String message = LogMessageUtil.buildMessage(LoggerService.errorMessages.get("INVALID_SDK_KEY_IN_OPTIONS"), null);
+            String message = "[ERROR]: VWO-SDK Please provide the sdkKey in the options and should be a of type string";
             System.err.println(message);
         }
 
         if (options == null || options.getAccountId() == null || options.getAccountId().toString().isEmpty()) {
-            String message = LogMessageUtil.buildMessage(LoggerService.errorMessages.get("INVALID_ACCOUNT_ID_IN_OPTIONS"), null);
+            String message = "[ERROR]: VWO-SDK Please provide VWO account ID in the options and should be a of type string|number";
             System.err.println(message);
         }
 
-        if (options.getIsAliasingEnabled() && (options.getGatewayService() == null || options.getGatewayService().isEmpty())) {
-            String message = LogMessageUtil.buildMessage(LoggerService.errorMessages.get("INVALID_GATEWAY_SERVICE_WHEN_ALIASING_ENABLED"), null);
+        if (options == null ||options.getIsAliasingEnabled() && (options.getGatewayService() == null || options.getGatewayService().isEmpty())) {
+            String message = "[ERROR]: VWO-SDK Please provide a valid gateway service url in the options when aliasing is enabled";
             System.err.println(message);
         }
         //start timer